DetectMun/web/requirements.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
